+++ tests/test_modeltest.py
@@ -1,5 +1,3 @@
-import sys
-
 import pytest
 
 from pytestqt.qt_compat import qt_api
@@ -113,33 +111,12 @@ def test_broken_types(check_model, broken_role):
     check_model(BrokenTypeModel(), should_pass=False)
 
 
-def check_broken_flag_or():
-    flag = qt_api.QtCore.Qt.AlignmentFlag
-    try:
-        int(flag.AlignHorizontal_Mask | flag.AlignVertical_Mask)
-    except SystemError:
-        # Should not be happening anywhere else
-        assert sys.version_info[:2] == (3, 11) and qt_api.pytest_qt_api == "pyside2"
-        return True
-    return False
-
-
-xfail_py311_pyside2 = pytest.mark.xfail(
-    check_broken_flag_or(),
-    reason="Fails to OR mask flags",
-)
-
-
 @pytest.mark.parametrize(
     "role_value, should_pass",
     [
-        pytest.param(
-            qt_api.QtCore.Qt.AlignmentFlag.AlignLeft, True, marks=xfail_py311_pyside2
-        ),
-        pytest.param(
-            qt_api.QtCore.Qt.AlignmentFlag.AlignRight, True, marks=xfail_py311_pyside2
-        ),
-        pytest.param(0xFFFFFF, False, marks=xfail_py311_pyside2),
+        pytest.param(qt_api.QtCore.Qt.AlignmentFlag.AlignLeft, True),
+        pytest.param(qt_api.QtCore.Qt.AlignmentFlag.AlignRight, True),
+        pytest.param(0xFFFFFF, False),
         ("foo", False),
         (object(), False),
     ],